assets/projects/nba_project/requirements.txt
streamlit
pandas
numpy
xgboost
scikit-learn
joblib
plotly
sqlalchemy
psycopg2-binary
uvicorn
fastapi
python-dotenv